rename package in service/controller/aws/v7patch1: v7 > v7patch1
