+++ service/controller/aws/v7patch1/error.go
@@ -1,4 +1,4 @@
-package v7
+package v7patch1
 
 import "github.com/giantswarm/microerror"
