+++ service/controller/aws/v7patch1/resource_set.go
@@ -1,4 +1,4 @@
-package v7
+package v7patch1
 
 import (
 	"context"
